perf: change image type from jpg to webp
initially the login page background image size was 5.3MB and was of very high resolution. so i reduce the size by changing its file type from jpg to webp
and also decreasing its resolution by 50%. it makes a significant improvement in file size of image which is now 433KB.

this increases the performance.

Fixes: #1296