Ser 2034/return clients for country (#103)
* feat: Add server-side filtering, pagination, and sorting to ClientsDataSource

- Update getClients method to include countryId parameter for filtering by country

* feat: Return groups and clients based on the selected country

* feat: Filter closed accounts in ClientsDataSource

This commit adds a filter to the ClientsDataSource class to exclude closed accounts from the list of clients. The filter is applied when the `showClosedAccounts` flag is set to false. This ensures that only active accounts are displayed to the user.

The code changes in this commit modify the `getClientsByCountry` and `getClients` methods to apply the filter. The `clients.pageItems` array is filtered to remove any clients with a status of "Closed".

* chore: Update @angular/cli to version 14.2.13 in Dockerfile

* feat: refactor subscription clients and group  to reduce duplication

* feat: Add country-based filtering for loan products

This commit modifies the LoanProductsResolver to include country-based filtering for loan products. It checks if a countryId is available in the session storage and retrieves loan products specific to that country. If no countryId is available, it retrieves all loan products as before.