fix: Make View Guarantors work even if no summary is sent by Fineract
This makes the View Guarantors work even if no Loan summary property is
sent in the Fineract response. (This is naturally the case when the
status of the Loan is not active.)
This also guarantees the page is rendered without errors in case
no clientRelationshipType is configured for the guarantor, which
is possible.

FIXES: WEB-22