build: update angular-eslint/schematics to latest version
Angular eslint 19.3 officially supports Angular from Angular 16 on,
but works well enough (only in special cases not) with Angular 14 and 15.
We therefore update the package to its latest version. This is the result
of "ng update @angular-eslint/schematics". There is a slight downgrade in
a peer dependency for eslint and typescript-eslint, but this is acceptable for
now.

Partially addresses WEB-73